Updatet all uint32_t sizes to size_t for larger strings/arrays on x64 architecture. Added StoreHelper for faster and easier storing of arrays. Added VarInt for sizes so most x64 binaries will be compatible with x32 binaries, plus smaller binary files. Added VarInt  tests.

+++ serialisation/include/serialisation/arrayInfo.h
@@ -28,16 +28,16 @@
 
 struct ArrayInfo
 {
-    uint32_t remainingCount;
+    size_t remainingCount;
     Internal::Type::Type type;
 
-    ArrayInfo( Internal::Type::Type type, uint32_t remainingCount )
+    ArrayInfo( Internal::Type::Type type, size_t remainingCount )
         : type( type ),
           remainingCount( remainingCount )
     {
     }
 
-    void Set( Internal::Type::Type type, uint32_t remainingCount )
+    void Set( Internal::Type::Type type, size_t remainingCount )
     {
         type = type;
         remainingCount = remainingCount;
